Move the code in the namespace modmesh from main.cpp to a header file
The classes moved are:

* REntityBase
* RStaticMesh
* RScene
