+++ src/viewer/main.cpp
@@ -26,194 +26,12 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <algorithm>
+#include <modmesh/modmesh.hpp>
+#include <modmesh/viewer/viewer.hpp>
 
 #include <QGuiApplication>
-#include <QByteArray>
-
-#include <Qt3DCore/QBuffer>
-#include <Qt3DCore/QEntity>
-#include <Qt3DCore/QGeometry>
-#include <Qt3DCore/QAttribute>
-#include <Qt3DCore/QTransform>
-
-#include <Qt3DRender/QCamera>
-
-#include <Qt3DExtras/QDiffuseSpecularMaterial>
-#include <Qt3DExtras/QCuboidMesh>
-
-#include <QOrbitCameraController>
 
 #include <qt3dwindow.h>
-
-#include <modmesh/modmesh.hpp>
-
-namespace modmesh
-{
-
-class REntityBase
-{
-
-public:
-
-    REntityBase(Qt3DCore::QEntity * ptr)
-        : m_ptr(ptr)
-    {
-    }
-
-    ~REntityBase()
-    {
-        if (nullptr != m_ptr)
-        {
-            // If not managed by a parent, signal for deletion.
-            if (nullptr == m_ptr->parent())
-            {
-                m_ptr->deleteLater();
-            }
-        }
-    }
-
-    void set(Qt3DCore::QEntity * ptr) { m_ptr = ptr; }
-
-    Qt3DCore::QEntity const * ptr() const { return m_ptr; }
-    Qt3DCore::QEntity * ptr() { return m_ptr; }
-
-    Qt3DCore::QEntity const * operator->() const { return m_ptr; }
-    Qt3DCore::QEntity * operator->() { return m_ptr; }
-
-private:
-
-    Qt3DCore::QEntity * m_ptr = nullptr;
-
-}; /* end class REntityBase */
-
-template <uint8_t ND>
-class RStaticMesh
-    : public REntityBase
-{
-
-public:
-
-    using mesh_type = StaticMesh<ND>;
-
-    RStaticMesh() = delete;
-    ~RStaticMesh() = default;
-
-    RStaticMesh(std::shared_ptr<mesh_type> const & static_mesh)
-        : REntityBase(new Qt3DCore::QEntity())
-        , m_static_mesh(static_mesh)
-        , m_geometry(make_geom(*static_mesh, ptr()))
-        , m_renderer(make_renderer(m_geometry))
-        , m_material(new Qt3DExtras::QDiffuseSpecularMaterial())
-    {
-        ptr()->addComponent(m_renderer);
-        ptr()->addComponent(m_material);
-    }
-
-    static Qt3DCore::QGeometry * make_geom(mesh_type const & mh, Qt3DCore::QEntity * root)
-    {
-        auto * geom = new Qt3DCore::QGeometry(root);
-
-        auto * buf = new Qt3DCore::QBuffer(geom);
-        {
-            QByteArray barray;
-            barray.resize(3 * mh.nnode() * sizeof(float));
-            float * ptr = reinterpret_cast<float *>(barray.data());
-            for (uint32_t ind = 0; ind < mh.nnode(); ++ind)
-            {
-                *ptr++ = mh.ndcrd(ind, 0);
-                *ptr++ = mh.ndcrd(ind, 1);
-                *ptr++ = 0;
-            }
-            buf->setData(barray);
-        }
-
-        auto * vertices = new Qt3DCore::QAttribute(geom);
-        vertices->setName(Qt3DCore::QAttribute::defaultPositionAttributeName());
-        vertices->setAttributeType(Qt3DCore::QAttribute::VertexAttribute);
-        vertices->setVertexBaseType(Qt3DCore::QAttribute::Float);
-        vertices->setVertexSize(3);
-        vertices->setBuffer(buf);
-        vertices->setByteStride(3 * sizeof(float));
-        vertices->setCount(mh.nnode());
-
-        geom->addAttribute(vertices);
-
-        uint32_t nend = 0;
-        for (uint32_t ifc = 0; ifc < mh.nface(); ++ifc)
-        {
-            nend += mh.fcnds(ifc, 0);
-        }
-
-        buf = new Qt3DCore::QBuffer(geom);
-        {
-            QByteArray barray;
-            barray.resize(nend * sizeof(uint32_t));
-            auto * indices = reinterpret_cast<uint32_t *>(barray.data());
-            for (uint32_t ifc = 0; ifc < mh.nface(); ++ifc)
-            {
-                for (int32_t inf = 1; inf <= mh.fcnds(ifc, 0); ++inf)
-                {
-                    *indices++ = mh.fcnds(ifc, inf);
-                }
-            }
-            buf->setData(barray);
-        }
-
-        auto * indices = new Qt3DCore::QAttribute(geom);
-        indices->setVertexBaseType(Qt3DCore::QAttribute::UnsignedInt);
-        indices->setAttributeType(Qt3DCore::QAttribute::IndexAttribute);
-        indices->setBuffer(buf);
-        indices->setCount(nend);
-
-        geom->addAttribute(indices);
-
-        return geom;
-    }
-
-    static Qt3DRender::QGeometryRenderer * make_renderer(Qt3DCore::QGeometry * geom)
-    {
-        auto * renderer = new Qt3DRender::QGeometryRenderer();
-        renderer->setGeometry(geom);
-        renderer->setPrimitiveType(Qt3DRender::QGeometryRenderer::Lines);
-        return renderer;
-    }
-
-    mesh_type const & static_mesh() const { return *m_static_mesh; }
-    mesh_type & static_mesh() { return *m_static_mesh; }
-
-private:
-
-    std::shared_ptr<mesh_type> m_static_mesh = nullptr;
-
-    Qt3DCore::QGeometry * m_geometry = nullptr;
-    Qt3DRender::QGeometryRenderer * m_renderer = nullptr;
-    Qt3DRender::QMaterial * m_material = nullptr;
-
-}; /* end class RStaticMesh */
-
-class RScene
-    : public REntityBase
-{
-
-public:
-
-    RScene()
-        : REntityBase(new Qt3DCore::QEntity())
-        , m_camera_controller(new Qt3DExtras::QOrbitCameraController(ptr()))
-    {
-    }
-
-    Qt3DExtras::QOrbitCameraController const * camera_controller() const { return m_camera_controller; }
-    Qt3DExtras::QOrbitCameraController * camera_controller() { return m_camera_controller; }
-
-private:
-
-    Qt3DExtras::QOrbitCameraController * m_camera_controller = nullptr;
-
-}; /* end class RScene */
-
-} /* end namespace modmesh */
 
 std::shared_ptr<modmesh::StaticMesh2d> make_3triangles()
 {
